requirements.txt
astroid==2.4.2
autocrop==1.1.0
autopep8==1.5.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
pycodestyle==2.6.0
pylint==2.6.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
